add many to many edit

+++ src/main/java/com/turbopick/autowise/repository/CarRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CarRepository extends JpaRepository<Car,Integer> {
-    public Car findCarById(int id);
+public interface CarRepository extends JpaRepository<Car,Long> {
+    public Car findCarById(Long id);
     public Car findCarByName(String name);
 
 }
